refactor: types/ を feat-recruit/common/types/ に移動
