+++ src/app/feat-recruit/common/types/recruit_condition_error.ts
@@ -1,4 +1,4 @@
-import { ErrorTexts } from '../../constant/error_texts';
+import { ErrorTexts } from '../../../constant/error_texts';
 
 export class RecruitConditionError extends Error {
     private errorMessage: string | undefined;
